Try to understand if the values fit the sequence boundaries.
When a negative margin is passed, there is a chance the sequence will not allow
the new value (common case: the resulting new value is negative).
Therefore, the function inspects the sequence definition and its boundaries
to understand if the new value can be safely set.
In the case it cannot be set, the function continues to the next sequence
after raising a warning to the user.